+++ tests/Query/FindCredentialsTest.php
@@ -5,12 +5,10 @@ namespace BSP\Credentials\Tests\Query;
 
 use BSP\Credentials\Entity\Credentials;
 use BSP\Credentials\Port\CredentialsReader;
-use BSP\Credentials\Port\PasswordEncoder;
 use BSP\Credentials\Query\FindCredentials;
 use BSP\Credentials\ValueObject\CredentialsId;
 use BSP\Credentials\ValueObject\Email;
 use BSP\Credentials\ValueObject\HashedPassword;
-use BSP\Credentials\ValueObject\PlainPassword;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -20,15 +18,13 @@ final class FindCredentialsTest extends TestCase
     {
         /** @var CredentialsReader|MockObject $credentialsReader */
         $credentialsReader = $this->createMock(CredentialsReader::class);
-        /** @var PasswordEncoder|MockObject $passwordEncoder */
-        $passwordEncoder = $this->createMock(PasswordEncoder::class);
         /** @var CredentialsId|MockObject $credentialsId */
         $credentialsId = $this->createMock(CredentialsId::class);
 
         $credentials = new Credentials(
             $credentialsId,
             new Email('john.snow@winterfell.north'),
-            new HashedPassword(new PlainPassword('winter is coming'), $passwordEncoder)
+            new HashedPassword('hashed-password')
         );
 
         $query = new FindCredentials($credentialsReader);
@@ -44,15 +40,13 @@ final class FindCredentialsTest extends TestCase
     {
         /** @var CredentialsReader|MockObject $credentialsReader */
         $credentialsReader = $this->createMock(CredentialsReader::class);
-        /** @var PasswordEncoder|MockObject $passwordEncoder */
-        $passwordEncoder = $this->createMock(PasswordEncoder::class);
         /** @var CredentialsId|MockObject $credentialsId */
         $credentialsId = $this->createMock(CredentialsId::class);
 
         $credentials = new Credentials(
             $credentialsId,
             new Email('john.snow@winterfell.north'),
-            new HashedPassword(new PlainPassword('winter is coming'), $passwordEncoder)
+            new HashedPassword('hashed-password')
         );
 
         $query = new FindCredentials($credentialsReader);